+++ NEWS.md
@@ -3,3 +3,4 @@
 ## Changes in version 0.97.0
 
 - Add *alabaster*-format stash for `XcmsExperiment`.
+- Add plain text-based format stash for `XcmsExperiment`.